refactor: CardSelectOption, Selector 이름, 역할을 범용적으로 변경
