+++ src/components/Option/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Dot } from '../Icon/assets';
+import './style.css';
+
+export default function Option({ name, color, ...props }) {
+  return (
+    <button type="button" className="option" aria-label={`${name} 선택 옵션`} {...props}>
+      <Dot className="option__button" size="36px" color={color} />
+      <p className="option__label">{name}</p>
+    </button>
+  );
+}
