+++ src/components/Select/style.css
@@ -1,4 +1,4 @@
-.card-company-selection {
+.select {
   display: grid;
   grid-template: repeat(2, 1fr) / repeat(4, 1fr);
   row-gap: 2rem;
