+++ src/components/index.js
@@ -1,8 +1,7 @@
-export { default as CardCompanySelection } from './CardCompanySelection';
-export { default as SecurityCodeGuide } from './SecurityCodeGuide';
+export { default as Select } from './Select';
 export { TextButton } from './Buttons';
 export { default as Card } from './Card';
-export { default as CardCompanyOption } from './CardCompanyOption';
+export { default as Option } from './Option';
 export { Input, CardPasswordInput, InputWithCounter } from './FormInputs';
 export { default as Icon } from './Icon';
 export { default as Modal } from './Modal';
