+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import '../styles/index.css';
+import './styles/index.css';
 import App from './App';
 
 ReactDOM.render(
