+++ src/styles/index.css
@@ -1,3 +1,4 @@
+@import './global.css';
 @import './shared/index.css';
 
 header {
