+++ src/styles/shared/index.css
@@ -1,4 +1,4 @@
-@import './global.css';
 @import './space.css';
 @import './background.css';
 @import './typography.css';
+@import './layout.css';
